Move operations interfaces into a dedicated directory
