+++ spec/Api/CurrencyApiSpec.php
@@ -4,7 +4,8 @@ namespace spec\Akeneo\Pim\ApiClient\Api;
 
 use Akeneo\Pim\ApiClient\Api\CurrencyApi;
 use Akeneo\Pim\ApiClient\Api\CurrencyApiInterface;
-use Akeneo\Pim\ApiClient\Api\ListableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\GettableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\ListableResourceInterface;
 use Akeneo\Pim\ApiClient\Client\ResourceClientInterface;
 use Akeneo\Pim\ApiClient\Pagination\PageFactoryInterface;
 use Akeneo\Pim\ApiClient\Pagination\PageInterface;
@@ -26,6 +27,7 @@ class CurrencyApiSpec extends ObjectBehavior
     {
         $this->shouldHaveType(CurrencyApi::class);
         $this->shouldImplement(CurrencyApiInterface::class);
+        $this->shouldImplement(GettableResourceInterface::class);
         $this->shouldImplement(ListableResourceInterface::class);
     }
