+++ src/Api/AssociationTypeApi.php
@@ -3,11 +3,9 @@
 namespace Akeneo\Pim\ApiClient\Api;
 
 use Akeneo\Pim\ApiClient\Client\ResourceClientInterface;
-use Akeneo\Pim\ApiClient\Exception\HttpException;
 use Akeneo\Pim\ApiClient\Exception\InvalidArgumentException;
 use Akeneo\Pim\ApiClient\Pagination\PageFactoryInterface;
 use Akeneo\Pim\ApiClient\Pagination\ResourceCursorFactoryInterface;
-use Psr\Http\Message\StreamInterface;
 
 /**
  * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
