+++ src/Api/ChannelApiInterface.php
@@ -2,6 +2,12 @@
 
 namespace Akeneo\Pim\ApiClient\Api;
 
+use Akeneo\Pim\ApiClient\Api\Operation\CreatableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\GettableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\ListableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\UpsertableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\UpsertableResourceListInterface;
+
 /**
  * API to manage the channels.
  *
