+++ src/Api/LocaleApiInterface.php
@@ -2,6 +2,9 @@
 
 namespace Akeneo\Pim\ApiClient\Api;
 
+use Akeneo\Pim\ApiClient\Api\Operation\GettableResourceInterface;
+use Akeneo\Pim\ApiClient\Api\Operation\ListableResourceInterface;
+
 /**
  * API to manage the locales.
  *
