+++ src/Api/Operation/DownloadableResourceInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Akeneo\Pim\ApiClient\Api;
+namespace Akeneo\Pim\ApiClient\Api\Operation;
 
 use Psr\Http\Message\StreamInterface;
